project/requirements.txt
fastapi==0.120.4
uvicorn==0.18.3
psycopg2-binary==2.9.10
pyjwt==2.8.0
pydantic
